refactor cache_helper into its own mod
